Modernize QodotMap property definitions
This one was a doozy. Formerly the QodotMap inspector interface was
populated by a `_get_property_list()` override. Now, the features that
_get_property_list provided are exposed through export annotations.

Also, the lack of typed arrays in Godot 3.5 required elaborate
workarounds in order to populate exported arrays with appropriate
contents. These workarounds are no longer necessary and have been removed.