Document a bunch of classes (#41)
* Document QodotPlugin

* Document QodotPlugin

* Document QodotUtil

* Document QodotEntity

* Document QodotFGDSolidClass

* Revert "Document QodotFGDSolidClass"

This reverts commit e7a35609e415e4be72486c31bb1b3c736bcd3d47.

* Document QodotNode3D (insofar as it can be)

* Modernize QodotMap property definitions

This one was a doozy. Formerly the QodotMap inspector interface was
populated by a `_get_property_list()` override. Now, the features that
_get_property_list provided are exposed through export annotations.

Also, the lack of typed arrays in Godot 3.5 required elaborate
workarounds in order to populate exported arrays with appropriate
contents. These workarounds are no longer necessary and have been removed.

* Document QodotMap properties

* Document QodotMap

---------

Co-authored-by: Blue Willow <blubug@protonmail.com>