Add support for Choices Default Value
Choices can have an optional default value passed in an FGD. Qodot was unable to do this before.

Proposed change has the optional default value defined in the Class Property Descriptions. To define it, use an Array with the format [String, int].

Falls back to normal behavior if the description is not for a Choices property or the description is not an Array.