Fix entity position bug (#80)
Fixed bug introduced by https://github.com/QodotPlugin/Qodot/pull/78 where it allowed positionless entities to be built without error, but an improper check caused all entities to not apply the 'origin' position.

Doing `node.get("position")` was not the proper way to handle the check. Fixed with the appropriate `"position" in node`.