Fixed broken Flags and Choices default properties (#79)
When building a map, if an entity is missing key value pairs due to not being filled out in Trenchbroom (grayed out key values) then Qodot will attempt to fill the missing properties with their defaults.

Flags and Choices are set up using Arrays and Dictionaries. If the value is found, they correctly apply integers to the property. If the value isn't found, they become the Arrays and Dictionaries defined in the FGD Class resource.

With the fix, Flags will now sum up properly to the default Flag settings, and Choices will automatically be set to the correct choice if the optional default is provided.