Added Entity Scale to TrenchbroomGameConfig Resource
The latest version of the TrenchBroom Game Config has support for game-wide entity model scaling. This change to Qodot's TrenchbroomGameConfig resource adds support for that feature.

The new variable is a String type that copies the String exactly as it's entered in the Resource Inspector. Doing it this way allows Qodot to support the multiple scale expression types that TrenchBroom supports.

This will prove extremely useful as users will no longer have to supply a scale for each individual model's meta properties or manually scale copies of each model for proper display in TrenchBroom.

See the [**TrenchBroom Docs**](https://trenchbroom.github.io/manual/latest/#game_configuration_files_entities) for more information on this feature.