Fix: Prevent Solid Entity Classes Having Size or Model Metaproperties (#94)
If a brush entity has a `size` meta property it will throw an error on FGD load.

> Reloading entity definitions
> Solid entity definition must not have a size (line 21, column 13)
> Loaded entity definition file QodotTutorial.fgd

The FGD will still load and you can still map, but an error's an error.

This fix adds a check to see if the resource is a Solid Class and if the current meta property is an invalid type for brush entities. If it is, it moves on to the next meta property.