Add Map Formats Array to TrenchbroomGameConfig
Small PR that allows users to specify what map formats they wish to support and the initial maps for them. Rather than use the default Trenchbroom map format array, I opted to default to Valve and the Quake formats, as Qodot currently only supports Quake and Valve UV formats, but a lot of work is being done to try to support both Quake 2 and Quake 3.

The Map Formats array in the GameConfig resource exports with respect to the order of the array's elements. To set a different default map format, simply reorder the array (or create a new one!).

The `initialmap` key value seems to be optional.

This should solve #102.