Fix: Connect Signal Assumes Node Will Have Properties
Qodot currently assumes that a Node will have a Properties dictionary when it sees some values populated for its corresponding Entity in a map file. Sometimes a `target` key value can get added to an entity through groups where it should not have one. Sometimes an entity is spawned with no scripts in place. This PR fixes it in the `connect_signal` function so that it no longer attempts to access a property that does not exist.